Simplify the test_runtime_dependencies_versions/1 test case
Vast simplifications are possible because there is no need to find
an Erlang system that can be started on the current machine. We are
only interested in the BEAM file in each release. Therefore, we
can bypass the test_server:find_release/1 function and directly
use the BEAM files in the directory where all OTP releases are
stored.