beam_disasm: Resolve {f,Label} Tag for call_fun2/3
In the `call_fun2/3` instruction, the Tag can either be `{f,Label}`,
`{atom,safe}` or `{atom,unsafe}`. When Tag is a known target (the
`{f,Label}`), `beam_asm:make_op/2` replaces the Tag with the index into
the lamba table.

This change looks up the index within the lambda table to resolve the
label.