beam_disasm: Fix translation of update_record/5
`update_record/5` needs some minor tweaks in order to be accepted by
`beam_validator`:

* `Size` must have its `{u,Size}` tag unwrapped
* The `Updates` list should discard its `z` and `u` tags and be placed
  in a `{list,List}` tuple.

This is similar to the translation in `beam_ssa_codegen:cg_instr/2`.
A similar treatment of the `list` change is seen in
`beam_disasm:resolve_inst/4` for the `init_yregs/2` instruction.

For an example listing:

    -record(r, {a, b}).
    update(#r{a = foo} = R) ->
        R#r{a = bar}.

we get the following asm:

    {test,is_tagged_tuple,{f,1},[{x,0},3,{atom,r}]}.
    {get_tuple_element,{x,0},1,{x,1}}.
    {test,is_eq_exact,{f,1},[{x,1},{atom,foo}]}.
    {test_heap,4,1}.
    {update_record,{atom,copy},3,{x,0},{x,0},{list,[2,{atom,bar}]}}.
    return.