Fix incorrect translation of low_watermark
low_watermark is a local inet_drv option, not to be confused
with SO_RCVLOWAT.  The inet_drv option works on the output buffer
to decide when to unblock a busy port, but the socket option
is an input threshold for when to deliver input data.

Not the same at all.

Incorrectly activating SO_RCVLOWAT on e.g a distribution socket
can block the initial distribution handshake and cause a failed
connection setup.