Don't allow local calls from record init in guards
Calling functions in guards is forbidden, but the compiler would
fail to reject a function call from a default record initializer
used in a guard. For example:

    -record(test, {a = mk_a()}).
    mk_a() -> 1.

    test_rec(Rec) when Rec =:= #test{} ->
        true.

Before this commit, the compiler would fail to reject the illegal call
to `mk_a/0`, and crash in a later pass. The compiler will now give the
following error message:

    t.erl:7:36: call to local/imported function mk_a/0 is illegal in guard
    %    7|         test_rec(Rec) when Rec =:= #test{} ->
    %     |                                    ^

Closes #6465
Closes #6466