kernel,pg: Do not send redundant sync 'messages'
I can't see how a second 'sync' message toward the same pg process
would be necessary at all.

When we get a 'discover' we monitor the sender and send a 'sync'
with all our local groups as reply. That's the only way a 'sync' is
sent. If we then get a second 'discover' from the same process, why
would that need a second 'sync' reply?

Any updates to our local groups are sent anyway as 'join' and 'leave'
messages.


pg_SUITE:forced_sync is just tweaked to not expect sending of
redundant 'sync'. With that change one would question its purpose,
or am I missing something?