Implement inet_epmd_dist
A distribution plug-in that implements the boilerplate
for distribution using EPMD over the inet address family.

inet6 is a future exercise.

inet_epmd_dist itself can take plug-ins that implement
the socket stuff, and it can itself act as a plug-in
to itself that mimics the default -proto_dist inet_tcp.

Also implement a plug-in to inet_epmd_dist: inet_epmd_socket
that implements the regular inet_tcp_dist distribution but
over the socket API.