[erts] Display large maps similarly to small maps
The formatting of large maps done by erlang:display/1 used to reveal
interesting information about the map internals, but was difficult to
read, compared to the formatting of small maps, and did not represent a
map in proper erlang syntax. This commit aims at a formatting that is
proper erlang syntax, still without any sorting of elements.