Reject invalid use of caret notation
It is documented that `$\^X` is the code point (ASCII code) for
Control X, where X is an uppercase or lowercase letter.

It turns out that it works for **all** characters:

    1> $\^@.
    0
    2> $\^_.
    31
    3> $\^Γ.
    19
    4> $\*.
    42
    5> $\^😀.
    0
    6> $\^?.
    31

Some of those are reasonable. `^@` and `^_` are allowed in Emacs and
most other tools that support the caret notation.

It is reasonable to allow `^?`, but the value of it should
be 127 (Delete) as in all other tools.

Allowing arbitrary letters and symbols does not make sense.

Therefore, this commit modifies the rules for the caret notation to
allow all characters in the range `16#40` through `16#4F` as well as
lowercase `a` through `z`. That is, the following characters are
allowed: `@`, `A`-`Z`, `[`, `\`, `]`, `^`, `_`, and `a`-`z`.  That
makes it possible to express all control codes from 0 through 31.

Also allowed is `?`, but the value of it will now be 127 instead
of 31. That is a potential incompatibility, but only for users who
used it despite it being undocumented.

Closes #6477