erts: Fix bug decoding unsorted flatmaps as keys in hashmap
"binary_to_term" of a hashmap (> 32 keys) with unsorted flatmap(s)
(<= 32 keys) as key(s) inside the hashmap.

The map term would be created but the map keys could not be found with
maps:get and friends. Other operations such as map compare and merge
could probably also give incorrect results.

This was only be a problem if the term was encoded outside the VM
by erl_interface, jinterface or otherwise, as the VM itself always
encodes flatmaps with sorted keys.