compiler: Implement SSA checking pass
This patch adds a new pass, enabled by `+{check_ssa,post_ssa_opt}`,
which runs after the ssa_opt pass. The pass uses SSA check directives
embedded as comments in the source file and parsed by erl_parse to
check the structure of the emitted BEAM SSA.

In the SSA checker, match rules are applied sequentially on the SSA
representation for a function, in the same order as if it had been
dumped to disk. If the current pattern doesn't match, the checker
tries with the next instruction. If the checker reaches the end of the
SSA representation without having matched all patterns, the check is
considered failed.

When a pattern is matched against an SSA instruction, the values of
variables already bound are considered and if the patterns matches,
free variables introduced in the successfully matched pattern are
bound to the values they have in the matched instruction.