Fix bug introduced in PR 6359
In https://github.com/erlang/otp/pull/6359/files, I used the new
ets:lookup_element/4 in various places in the standard library.

In particular, I replaced a try/catch wrapping a call to
lookup_element/3 in 3 places. Unfortunately, this catch was also
catching the case where the table does not exist at all, and so now
things are breaking.

This patch just restores the 3 try/catches.