multisend
This commit implements an initial multisend implementation. The issue here is
that it changes the implementation of `send` to be implemented in Erlang instead
of directly in C, which incurs a penalty on all regular and existing uses of
`send`.

This commit originally included changing many preloaded binaries but this have
been removed from the commit history in order to comply with the security
requirements of PRs to OTP.