kernel: Accept ver5 distribution connection setup
We accept old 'send_name' handshake messages used by OTP-22 and older.
OTP-23 or OTP-24 nodes, not using epmd, may send the old message
to us if they do not know our OTP version.
Hence, we accept the old 'send_name' and the accompanying
'send_complement', but we do not send them ourself.

This was removed prematurely in OTP-25.0 by
commit 722219f211092e0c4b919f3ba4f91df42191c0e7.
Therefore versions 25.0 to 25.2.* are broken in this regard and may
reject connection setup from OTP-23 and 24 not using epmd.