erts: `recv_marker_reserve` needs stack too
I was a bit too quick in merging the fix for this bug and forgot
that HAlloc requires the stack pointer to be updated as well. This
commit should not actually make a difference since the bug only
occurred on 32-bit systems where the JIT isn't used, but it's better
to be consistent throughout.