For clarity, introduce new members in enum Update
Instead of having to write:

    emit_enter_runtime<Update::eStack | Update::eHeap>();

when we are about to call a function that allocates using the
`HAlloc()` function, introduce `Update::eHeapAlloc` so we can
write:

    emit_enter_runtime<Update::eHeapAlloc>();

Introduce `Update::eHeapOnlyAlloc` to be used when we are about
to call a function that allocates using `HeapOnlyAlloc()`:

    emit_enter_runtime<Update::eHeapOnlyAlloc>();

In a release build, `Update::eHeapOnlyAlloc` only causes the heap
pointer to be saved. In a debug build, both the heap pointer and
the stack pointer will be saved to ensure that the heap overflow
assertions will work.