Do not allocate a new map when the value is the same encore
This is a followup to PR #1889 almost 5 years in the making. The optimisation there
was implemented for many cases, but somehow `M#{existing_key => same_value}`
and `M#{existing_key => new_value}` cases were missed.

This is a bit more complicated since there are two cases to handle:
* the keys and values didn't change, and we can just return the original map
* the keys didn't change, but values did - we return a new map but keeping original keys tuple

This seems to be hit farily frequently - while I was working on it, it wasn't even possible to
start the VM successfully.