Unlink port_command to avoid EXIT msgs
A {'EXIT', Port, normal} message is always sent after
the port is closed.

If the user traps exit signal it will end up in the msg queue.
Unlink the port to avoid that.

If the program doesn't not exist open_port will exit directly,
so catch that to improve error handling, also redirect
stderr_to_stdout to catch any error messages from the command.

Update testcase to catch error on mac