Replace size/1 with either tuple_size/1 or byte_size/1
The `size/1` BIF is not optimized by the JIT, and its use can result
in worse types for Dialyzer.

When one knows that the value being tested must be a tuple,
`tuple_size/1` should always be preferred.

When one knows that the value being tested must be a binary,
`byte_size/1` should be preferred. However, `byte_size/1` also accepts
a bitstring (rounding up size to a whole number of bytes), so one must
make sure that the call to `byte_size/1` is preceded by a call to
`is_binary/1` to ensure that bitstrings are rejected. Note that the
compiler removes redundant calls to `is_binary/1`, so if one is not
sure whether previous code had made sure that the argument is a
binary, it does not harm to add an `is_binary/1` test immediately
before the call to `byte_size/1`.