base64: allow skipping padding = characters
This extends base64 encoding and decoding routines to allow
skipping emitting final = characters, or to accept encoded
data with final = characters skipped. This format is common
for some use-cases. Today, parsing it requires appending
the missing = characters, which incurs an unnecessary copy
of the input data, or re-implementing the entire routine.

This amends the interface in a backwards-incompatible way,
since the last change wasn't released yet, so should have
no expectation of backwards-compatibility.

Additionally, tests are slightly enhanced to cover more cases
directly with simple examples.