Remove the file_io_servers ETS table
This table was used to store Pids of files opened through the file server,
however this data was never used by OTP itself, except for the debugging
file:pid2file/1 function. Given the function is deprecated, we can skip the
overhead of the central ETS table, and store the name in the file process
itself. This makes the pid2file function more expensive, but as a deprecated
debugging facility this is acceptable given reduction in overall
memory overhead and synchronisation.