jit: Refactor breakpoints
The ARM reference manual, section B2.2.5, says that it's unsafe
to concurrently modify non-branch instructions. The hardware has
license to explode in our face even if we don't care whether the
new or old instruction executes.

To be safe, change the MOVZ in the breakpoint to a (slightly
slower) LDR in the fragment, letting us change the flag without
altering any instructions.

While we're at it we'll give x86 the same treatment. The Intel
manual is rather unclear about the subject, sometimes saying that
we don't need to do anything special[0] and sometimes that we
do[1]. I've decided to play it safe by only modifying the initial
jump, as there aren't as many things that could go wrong with that.

[0]: Intel 64/IA32 software dev manual Vol 3A, section 11.6
[1]: Intel 64/IA32 software dev manual Vol 3A, section 8.1.3