Deprecate file:pid2name/1
The `file:pid2name/1` function is limited in that it does not work for
files opened in `raw` mode or for files opened on a remote file
server. The implementation also adds overhead for each open and close
operation because an ETS table must be updated.

Therefore, deprecate the `file:pid2name/1` now, and schedule it for
removal in Erlang/OTP 27.