jit: Clear instruction cache ourselves
`__builtin__clear_cache` issues full memory and instruction
barriers in addition to clearing the cache. This is bad because it
may hide bugs where we've failed to issue barriers on all
schedulers before executing code.

If we clear the cache ourselves without issuing these barriers,
we'll crash really quick if we forget to issue them for all
schedulers later on.