maps: raise a badmap error if an iterator fails later
The implementation of filter/2, filtermap/2, fold/3, foreach/2 and map/2
raises a badmap exception if and only if the given iterator is not an
iterator at all, ie when the first call to next/1 on it fails. If a later
subsequent call to next/1 fails (ie, if something like {x, y, z} is given
as an iterator), the result will be a badarg error instead.

The changes in this commit change this behavior so that a badmap error is
always raised in such circumstances, that is, even if the given iterator
fails only on later subsequent calls.

Likewise, the stdlib error reporting did not complain about an iterator
being faulty if the first call to maps:next/1 on it succeeded. This has
also been addressed by the changes in this commit.