Close spawn driver write file descriptor (EOF.)
This implements a means to close the stdin of a port (send EOF). This is
necessary for piping data to programs which usually read from a file but
can read from stdin.

It works by closing the write descriptor by means of sending a magic
code to port_control/3. The write descriptor is replaced by /dev/null,
with further writes being discarded to the bit bucket.

Further reading:
  - https://erlang.org/pipermail/erlang-questions/2013-July/074916.html
  - https://github.com/erlang/otp/issues/4326

Example usage:
    erlang:port_control(Port, 16#0112c000, []).