STDLIB tests: Fix falling map iterator tests in debug build
The `>=/2` sort function is not appropriate for sorting maps with
mixed integer and float keys. It happens to work with small maps
because `maps:to_list/1` returns a list already correctly sorted. In a
debug build, the four-elements maps are large maps and the test case
fails.

While at it, also correct `ARevCmpFun` in `maps_SUITE`.