compiler: Don't generate unsafe code for binary append
The private-append optimization worked on the assumption that the
`private_append` version of `bs_create_bin` checked that the type of
the first argument was a bitstring. As it doesn't, the compiler
shouldn't do the optimization when it cannot guarantee that the type
is and only is a bitstring.

Fixes #6851
Fixes #6848