compiler: Avoid crash in private-append optimization
Avoid a crash in the private-append optimization due to attempting to
rewrite a literal which isn't a bitstring.

As value tracking is done without type information, it can happen that
when following def chains, we encounter literals which are not
bitstrings. We should never schedule a literal which isn't a bitstring
for rewriting when the complete literal is what we are tracking,
i.e. when the element is `self`.

Closes #6847