Improve the lists:merge family of functions
The merge functions no longer accept non-list arguments for merging.
Before, non-list arguments were accepted and returned if they were
the first argument and the second (and third, where applicable,
ie in merge3/3 and umerge3/3) arguments were empty lists. For merge/1
and umerge/1, non-list arguments were accepted if they were the only
element in a list of otherwise empty lists.

The merge functions now also return the second (or third, where
applicable) list immediately and unchanged if the other lists
are empty.
Before, this was only done for the first argument, if the second
(and third, where applicable) list was empty. In all other cases,
the merging process would be performed, unnecessarily.

The same applies for the undocumented reverse-merge functions like
rmerge/2,3, rumerge/2,3 etc.

Co-authored-by: Jan Uhlig <juhlig@hnc-agency.org>