Eliminate "mutable variables"
The compiler would generate incorrect code for the
following type of expression:

    Pattern = BoundVar1 = . . . = BoundVarN = Expression

An exception should be raised if any of the bound variables have
different values than `Expression`. The compiler would generate code
that would cause the bound variables to be bound to the value of
`Expression` whether the value matched or not.

For example:

    t() ->
        Zero = 0,
        One = 1,
        Result = One = Zero,
        {Result,One,Zero}.

There should be a `badmatch` exception, but instead, the `t/0`
function returned `{0,0,0}`.

Closes #6873