Add best practice for atoms in NIFs
Mention that atoms created in load are available in all environments.
The best practice is to pre-create all of them before usage, make sure
this best practice is included in the documentation.

Also fixed a "has" that should have been a "have".

Use full example for pre-created atoms and use _atom suffix for variable
names in the example (atom_ prefix would make variable name expansion in
IDEs less than useful).

Co-authored-by: Sverker Eriksson <sverker@erlang.org>