erl_lint: fix unbound singleton types in unions
Dialyzer (erl_lint) was accepting code of the following form:

```erlang
-spec run_test(Opts) -> term()
      when Opts :: {join_specs, Bool} | {test, Bool}.
run_test(Opts) ->    {error, fail}.
```

But this cannot be ever checked by Dialyzer as the type variable `Bool`
is in an union, which means that `Bool` happened only once and thus, `Bool`
is an unbound type variable. Essentially, the example above should be
rejected by Dialyzer in the same manner as this other one below:

```erlang
-spec run_test_error(Opts) -> term()
      when Opts :: {join_specs, Bool}.
run_test_error(Opts) ->
    {error, fail}.
```

This commit fixes GH-6508 and adds a bunch of tests to `erl_lint.erl`
so that we can keep track of which singleton type variables should throw
an error and which ones are ok.