public_key: Move decode of CRLDistributionPoints extension
As different solutions of verifying certificate revocation exists
move the decode of 'CRLDistributionPoints' so that it will only be decode
when it is actually used in the verification process. This would enable
interoperability with systems that use certificates with an invalid empty
CRLDistributionPoints extension that they want to ignore and make verification
by other means.

Closes #6402