ftp: create internal typed api
removes some type information from module `ftp.erl` because this
information should not be needed by users of `ftp`.

before this change, users of `ftp` module knew that if they get an error
as follows `{error, Reason}`, `Reason` was a specific atom(), i.e.,
`Reason :: 'ehost' | ...`. users of `ftp` should not rely on the
internal atom used, since they are expected to call `ftp:formaterror/1` to
understand the reason behind the error, as in any other OTP app.

however, this type information can be useful to maintainers of OTP.
Thus, `ftp.erl` contains the necessary type information for users of the
ftp application. the new module `ftp_internal.erl` has been introduced and
contains the actual implementation of the functions from `ftp.erl`, so that
`ftp.erl` is now a simple wrapper around `ftp_internal.erl`.

`ftp_internal.erl` contains the more precise error types for the type
variable `Reason` and also contains the implementation of the `ftp`
functions.