Ensure that the bs_match instruction can be extended
If we were to introduce new sub commands for the `bs_match`
instruction in the future, we want to make sure that loading a BEAM
file produced by a future Erlang/OTP version will not crash the
runtime when loaded into an earlier version. Therefore, fail loading
gracefully if any unknown sub commands are found. For example, if we
were to introduce support for matching floats in a future release,
loading would fail like so if we try to load that module in Erlang/OTP
26:

    =ERROR REPORT==== 20-Feb-2023::13:03:21.664064 ===
    beam/beam_load.c(583): Error loading function bad_bs_match:f/1:
      bs_match: invalid sub instruction 'float' (is this BEAM file generated by a future version of Erlang/OTP?)