compiler: Traverse call graph breadth first during alias analysis
During alias analysis the main data flow of alias information is from
caller to callee when the aliasing status of function arguments are
transferred. A smaller flow goes from the callee to the caller with
the information about the returned value.

As up to date alias status is needed at call sites (derived from the
status of the caller's own arguments) and the status recorded at all
callsites is what allows the analysis of the callee to provide a
status for the returned result in the caller, convergence of the alias
analysis can be greatly sped up by traversing the call graph breadth
first.