Eliminate a valgrind warning
When matching an `utf8` segment, valgrind could generate a warning
stating that a jump or conditional move was based on an uninitialized
value. The following instructions were the culprit:

    test rax, rax
    short jns L107

Those instructions test the most significant bit of the `rax` register
and jumps if it is zero. The most significant bit is always
initialized when the `test` instruction is executed, but it is
possible that some of the lower bits in the register are
uninitialized. It seems that valgrind will emit the warning if any
part of the register is uninitialized when executing the `test`
instruction (presumably because the values of the flags `ZF` and `PF`
depend on all bits in the register including the uninitialized bits).

We can avoid this warning by using the `bt` instruction, which only
sets the CF flag:

    bt rax, 63
    short jnc L107