Make add_transport start failure more synchronous
If diameter_service:start_transport/2 fails return a start failure
to the diameter_config child start to avoid that the supervisor
thinks the child has started and registers it for a brief moment
only to get a link message due to the start failure.

This may change an `{error,Reason}` to a `Reason` indicating
a supervisor child start error.