Correct unsafe optimization of binary matching
When compiling generated code from forms where the line numbers in the
annotations did not correspond to the actual order of the clauses, the
compiler could rearrange clauses even when it was not safe. For
example:

    match_route([<<"prefix">>, <<"action">>]) -> first;  %Line 4
    match_route([<<"prefix">>, _Ignore]) -> second.      %Line 2

The compiler would rearrange that to:

    match_route([<<"prefix">>, _Ignore]) -> second;      %Line 2
    match_route([<<"prefix">>, <<"action">>]) -> first.  %Line 4

which would make the second clause (former first clause) unreachable.

This bug was introduced in cadd19e84478ed (#6538).

Closes #6923