compiler: Avoid unsafe rewrite to bs_create_bin private_append
Nick Vatamaniuc reported that the following Erlang snippet:

```
B1 = << <<"x">> || true >>,
B2 = <<B1/binary, "z">>,
```

created identical `B1` and `B2` binaries. The error occurs because the
private-append pass was missing the condition that for a transform
from `bs_create_bin append, ...` to `bs_create_bin private_append,
...` to be safe, the first fragment must die with the `bs_create_bin`
instruction, it is not enough that the fragment is unique and
appendable.

The error is avoided by having the alias analysis, which has access to
liveness information, annotate `bs_create_bin` instructions with a
boolean annotation indicating whether the first fragment dies with the
instruction. The later private-append pass can then consult the
annotation to determine if the transform is safe. The use of an
annotation is more efficient than trying to reconstruct the
information in the alias analysis kill map during the private-append
pass.

Fixes #6925