Make maps_order optional in io_lib:format_spec()
The 'maps_order' key has been added to the public type
io_lib:format_spec() in OTP26 RC1, as a required key.
Besides, the key presence is assumed in the implementation
of io_lib:unscan_format/1. This makes it a breaking change,
resulting in unexpected behavior for existing code calling
it on a map without the key (returns a list of maps).

- fix the typespec to make the key optional

- don't require the key in io_lib:unscan_format/1

- add a guard to detect we're not building a correct charlist