erts: Reintroduce literal fun optimization
The optimization that turned make_fun2/3 instructions of funs
lacking an environment into literals was lost during the pre-JIT
refactoring of the loader, as creation of fun entries was now
(rightly) deferred until the very last stages of module
loading.

This commit restores that optimization by creating placeholders
in the form of external funs, which are then converted to local
funs during the last stages of module loading.